requirements.txt
newspaper3k==0.2.8
nltk==3.8.1
Requests==2.31.0
